+++ asn1_parser/constr_TYPE.h
@@ -30,7 +30,7 @@ struct asn_TYPE_member_s;	/* Forward declaration */
 typedef struct asn_struct_ctx_s {
 	short phase;		/* Decoding phase */
 	short step;		/* Elementary step of a phase */
-	int context;		/* Other context information */
+	size_t context;		/* Other context information */
 	void *ptr;		/* Decoder-specific stuff (stack elements) */
 	ber_tlv_len_t left;	/* Number of bytes left, -1 for indefinite */
 } asn_struct_ctx_t;
